Lemon API: Basic feedback endpoints created

+++ lemon.go
@@ -7,10 +7,10 @@ type Lemon struct {
 }
 
 type Feedback struct {
-	ID int `json:"id" db:"id"`
-	Rating int `json:"rating" db:"rating"`
-	Description string `json:"description" db:"description"`
-	Type string `json:"type" db:"type"`
-	Submitted *time.Time `json:"submitted" db:"submitted"`
-	Read bool `json:"read" db:"read"`
+	ID          int64      `json:"id" db:"id"`
+	Rating      int64      `json:"rating" db:"rating"`
+	Description string     `json:"description" db:"description"`
+	Type        string     `json:"type" db:"type"`
+	Submitted   *time.Time `json:"submitted" db:"submitted"`
+	Read        bool       `json:"read" db:"read"`
 }
